refactor: profile 이미지를 보여주는 section 분리
