+++ client/src/styles/styled.d.ts
@@ -14,7 +14,8 @@ declare module 'styled-components' {
       purple: string;
       blue: string;
       green: string;
-      lightGreen: string;
+      greenLight: string;
+      greenLightActive: string;
       greenOpactiy30: string;
       grayLight: string;
       grayLightActive: string;
